refactor: extract synthetic benchmark catalog
