+++ devtools/benchmark_campaigns.py
@@ -11,7 +11,7 @@ from dataclasses import dataclass, field
 from datetime import UTC, datetime
 from pathlib import Path
 
-from polylogue.scenarios import ScenarioMetadata
+from .synthetic_benchmark_catalog import SYNTHETIC_BENCHMARK_REGISTRY, SYNTHETIC_BENCHMARK_SCENARIOS
 
 
 @dataclass
@@ -27,84 +27,6 @@ class CampaignResult:
     artifact_targets: list[str] = field(default_factory=list)
     operation_targets: list[str] = field(default_factory=list)
     tags: list[str] = field(default_factory=list)
-
-
-@dataclass(frozen=True)
-class SyntheticBenchmarkScenario(ScenarioMetadata):
-    """Authored scenario metadata for synthetic long-haul benchmark campaigns."""
-
-    scenario_id: str
-    description: str
-    summary_metric: str
-    summary_label: str
-
-
-SYNTHETIC_BENCHMARK_SCENARIOS: tuple[SyntheticBenchmarkScenario, ...] = (
-    SyntheticBenchmarkScenario(
-        scenario_id="fts-rebuild",
-        description="Benchmark full FTS5 index rebuild",
-        summary_metric="rebuild_wall_s",
-        summary_label="s",
-        origin="authored.synthetic-benchmark",
-        artifact_targets=("message_fts",),
-        operation_targets=("index.message-fts-rebuild",),
-        tags=("benchmark", "synthetic", "fts"),
-    ),
-    SyntheticBenchmarkScenario(
-        scenario_id="incremental-index",
-        description="Benchmark incremental FTS index updates",
-        summary_metric="total_wall_s",
-        summary_label="s",
-        origin="authored.synthetic-benchmark",
-        artifact_targets=("message_fts",),
-        operation_targets=("index.message-fts-incremental",),
-        tags=("benchmark", "synthetic", "fts"),
-    ),
-    SyntheticBenchmarkScenario(
-        scenario_id="filter-scan",
-        description="Benchmark common filter query patterns",
-        summary_metric="list_50_wall_s",
-        summary_label="s",
-        origin="authored.synthetic-benchmark",
-        artifact_targets=("conversation_query_results",),
-        operation_targets=("query.filters.synthetic-scan",),
-        tags=("benchmark", "synthetic", "filters"),
-    ),
-    SyntheticBenchmarkScenario(
-        scenario_id="startup-health",
-        description="Benchmark check --runtime startup speed",
-        summary_metric="total_health_s",
-        summary_label="s",
-        origin="authored.synthetic-benchmark",
-        artifact_targets=("archive_health",),
-        operation_targets=("health.startup.synthetic",),
-        tags=("benchmark", "synthetic", "health"),
-    ),
-    SyntheticBenchmarkScenario(
-        scenario_id="action-event-materialization",
-        description="Benchmark action-event read-model rebuild over synthetic tool-use transcripts",
-        summary_metric="rebuild_wall_s",
-        summary_label="s",
-        origin="authored.synthetic-benchmark",
-        artifact_targets=("tool_use_source_blocks", "action_event_rows", "action_event_fts"),
-        operation_targets=("materialize-action-events",),
-        tags=("benchmark", "synthetic", "action-events"),
-    ),
-    SyntheticBenchmarkScenario(
-        scenario_id="session-product-materialization",
-        description="Benchmark durable session-product rebuild over synthetic archive conversations",
-        summary_metric="rebuild_wall_s",
-        summary_label="s",
-        origin="authored.synthetic-benchmark",
-        artifact_targets=("session_product_source_conversations", "session_product_rows", "session_product_fts"),
-        operation_targets=("materialize-session-products",),
-        tags=("benchmark", "synthetic", "session-products"),
-    ),
-)
-
-SYNTHETIC_BENCHMARK_REGISTRY: dict[str, SyntheticBenchmarkScenario] = {
-    scenario.scenario_id: scenario for scenario in SYNTHETIC_BENCHMARK_SCENARIOS
-}
 
 
 def _db_row_counts(db_path: Path) -> dict[str, int]:
@@ -550,7 +472,6 @@ __all__ = [
     "CampaignResult",
     "SYNTHETIC_BENCHMARK_REGISTRY",
     "SYNTHETIC_BENCHMARK_SCENARIOS",
-    "SyntheticBenchmarkScenario",
     "run_filter_scan_campaign",
     "run_fts_rebuild_campaign",
     "run_action_event_materialization_campaign",
