+++ devtools/benchmark_catalog.py
@@ -7,7 +7,7 @@ from dataclasses import dataclass
 from polylogue.scenarios import ScenarioMetadata
 
 from .benchmark_campaign import CAMPAIGNS as BENCHMARK_CAMPAIGNS
-from .benchmark_campaigns import SYNTHETIC_BENCHMARK_SCENARIOS
+from .synthetic_benchmark_catalog import SYNTHETIC_BENCHMARK_SCENARIOS
 
 
 @dataclass(frozen=True)
