+++ devtools/scenario_projection_catalog.py
@@ -9,7 +9,7 @@ from polylogue.scenarios import (
 from polylogue.showcase.exercises import EXERCISES
 
 from .benchmark_campaign import CAMPAIGNS as BENCHMARK_CAMPAIGNS
-from .benchmark_campaigns import SYNTHETIC_BENCHMARK_SCENARIOS
+from .synthetic_benchmark_catalog import SYNTHETIC_BENCHMARK_SCENARIOS
 
 
 def build_scenario_projection_entries() -> tuple[ScenarioProjectionEntry, ...]:
